refactor(openclaw-config): extract pdf/image model resolution from build.ts
Move resolveDefaultPdfModel/resolveDefaultImageModel, their explicit provider
preference tables, and OLLAMA_CLOUD_IMAGE_PREFERENCE into default-media-models.ts
— a self-contained 'pick the vision model for the built-in pdf/image tools'
concern. build.ts imports the two resolvers; its model-vision and
model-capability-cache imports drop with the code that used them.

The #416 drift guard now imports the preference list from its new home. No
behaviour change; openclaw-config + drift suites (191 tests) stay green.
